\name{SpecData}
\alias{SpecData}
\alias{MetaDataDDT}
\alias{MetaDataKelthane}
\docType{data}
\title{Package data.}
\description{Metadata and spectral data for SpecLibDDT2015. Called by \code{LibraryReport()}.}
\usage{SpecData}
\format{\code{SpecData} is a data frame containing the mass spectra.  \code{MetaDataDDT} and \code{MetaDataKelthane} are data frames containing the metadata for each of the technical mixtures.
}
\keyword{datasets} 
