\name{SpecLibDDT2015-package}
\alias{SpecLibDDT2015-package}
\docType{package}
\title{
Mass Spectral Library for DDT and Kelthane-Dicofol Technical Mixtures
}
\description{
Contains the code and mass spectral library described in: TODO

The corresponding article includes a description of
the study goals, samples, and GCxGC/TOF-MS
analytical method. The original PDF version of this library is supplementary
material to the article.
}
\keyword{ package }
